// src/vs/platform/agentHost/node/codex/protocol/generated/v2/SandboxPolicy.ts
// AUTOGENERATED - do not edit.
// Source: `codex app-server generate-ts --experimental`
// Generated from @openai/codex 0.146.0, licensed Apache-2.0.
// Modified by VS Code to rewrite imports and apply repository formatting.
// See README.md in this directory for provenance and licensing notes.
// Regenerate: npm run codex:gen-protocol
// Source file: v2/SandboxPolicy.ts

import type { AbsolutePathBuf } from "../AbsolutePathBuf.js";
import type { NetworkAccess } from "./NetworkAccess.js";

export type SandboxPolicy = { "type": "dangerFullAccess" } | { "type": "readOnly", networkAccess: boolean, } | { "type": "externalSandbox", networkAccess: NetworkAccess, } | { "type": "workspaceWrite", writableRoots: Array<AbsolutePathBuf>, networkAccess: boolean, excludeTmpdirEnvVar: boolean, excludeSlashTmp: boolean, };